% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{cloudmask}
\alias{cloudmask}
\title{Compute MODIS cloudmask}
\usage{
cloudmask(x)
}
\arguments{
\item{x}{numeric. MODIS QA values.}
}
\value{
A vector with the same dimensions as \code{x}, taking the values \code{1} and
\code{NA} for clear and cloudy pixels, respectively.
}
\description{
Compute MODIS cloudmask
}
