% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{combine}
\alias{combine}
\title{Combine elements}
\usage{
combine(...)
}
\arguments{
\item{...}{arguments}
}
\value{
A \code{list} of all possible combinations.
}
\description{
Create all unique element combinations.
}
\examples{
\dontrun{
combine(x = 1:3, y = letters[1:2])
}
}
