% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\alias{compose,Database,Cropmaps-method}
\title{Compose rasters by maps}
\usage{
compose(x, y, ...)

\S4method{compose}{Database,Cropmaps}(x, y, variablex, variabley, fun = "mean", ...)
}
\arguments{
\item{x}{S4 object. The product of interest.}

\item{y}{S4 object. The product determining the zones.}

\item{...}{extra arguments passed to \code{fun}.}

\item{variablex}{character. The variable of product \code{x}.}

\item{variabley}{character. The variable of product \code{y}.}

\item{fun}{character or function. The function to apply for the summaries.}
}
\value{
A data.frame with the zonal summaries. The data.frame is also saved in the cronus database in .csv format.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Compose Rasters into zonal summaries, indicated by another raster.
}
\details{
Currently, \code{y} can only be a NASS Cropmap.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

# Compute the Growing Degree Days using Daymet and CDL.
x <- new("Daymet", region = region, date = date, dir = dir)
y <- new("Cropmaps", region = region, date = date, dir = dir)
a <- compose(x, y, variablex = "gdd", variabley = "cdl_default", fun = "mean")
}
}
