% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congregate.R
\name{congregate}
\alias{congregate}
\alias{congregate,Cropmaps-method}
\title{Congregate raster data}
\usage{
congregate(x, ...)

\S4method{congregate}{Cropmaps}(x, variable)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. A function to compute the variable of interest.}
}
\value{
data.frame. A data frame with the congregated data.
}
\description{
Congregate raster data into a data frame.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

## Cropmaps CDL

# Create the object
x <- new("Cropmaps", region = region, date = date)

# congregate
a <- congregate(x, "cdl_recoded")
}
}
