% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{create_keyring}
\alias{create_keyring}
\alias{delete_keyring}
\alias{log_in}
\alias{log_out}
\alias{add_key}
\alias{get_username}
\alias{get_password}
\alias{delete_key}
\alias{check_keyring}
\title{Manage keys and keyrings}
\usage{
create_keyring(ringname, password)

delete_keyring(ringname)

log_in(ringname, password)

log_out(ringname)

add_key(ringname, provider, username, password)

get_username(ringname, provider)

get_password(ringname, provider)

delete_key(ringname, provider)

check_keyring(ringname)
}
\arguments{
\item{ringname}{character. The name of a keyring (a ringname).}

\item{password}{character. A password.}

\item{provider}{character. The name of a provider (e.g. \code{"modis"}).}

\item{username}{character. A username signed-up in a service provider.}
}
\value{
nothing.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This set of functions handles keys that are needed to download products such
as satellite data.
}
\details{
This set of function are wrappers of the keyring package. A keyring
is an account that can hold keys (credentials) for various services. Once a
keyring is created, the user can start adding keys to it. Functions in the
\code{cronus} package that require credentials can automatically get them if
provided with the ringname. Users are reminded to log-in and log-out
in order to use their keyrings.
}
\section{Functions}{
\itemize{
\item \code{create_keyring()}: Create a keyring.

\item \code{delete_keyring()}: Delete a keyring.

\item \code{log_in()}: Log into your keyring.

\item \code{log_out()}: Log out of your keyring.

\item \code{add_key()}: Add a key to your keyring.

\item \code{get_username()}: Get a username.

\item \code{get_password()}: Get a password.

\item \code{delete_key()}: Delete a key.

\item \code{check_keyring()}: Check whether the user is logged-in.

}}
\examples{
\dontrun{
ringname <- "my_ringname"
password <- "my_password"

create_keyring(ringname, password)
check_keyring(ringname)

log_in(ringname, password)
check_keyring(ringname)

add_key(ringname = ringname,
provider = "nass",
username = NULL,
password = "nass_key")

add_key(ringname = ringname,
        provider = "usgs",
        username = "usgs_username",
        password = "usgs_password")

get_username(ringname, "usgs")
get_password(ringname, "usgs")
delete_key(ringname, "usgs")

log_out(ringname)
delete_keyring(ringname)
}
}
