% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{create_rtoi}
\alias{create_rtoi}
\title{Create an rtoi object}
\usage{
create_rtoi(region)
}
\arguments{
\item{region}{character. The region of interest.}
}
\value{
A data.frame with the requested data.
}
\description{
This function creates an rtoi object for a specific region, according to
the rsat package.
}
\details{
\code{create_rtoi()} is a wrapper function, calling the \code{new_rtoi()} function from
the \code{rsat} package. Both the \code{db_path} and \code{rtoi_path} are created inside a
temporary directory, using \code{tempdir()}.
}
\examples{
\dontrun{
create_rtoi("nebraska")
}
}
\seealso{
\code{\link[rsat:new_rtoi]{rsat::new_rtoi()}}
}
