% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronus-package.R
\docType{package}
\name{cronus-package}
\alias{cronus}
\alias{cronus-package}
\title{cronus: Agricultural Remote Sensing Database Constructor}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

cronus is an R package that focuses on database creation. It can be used to download and process agricultural, environmental and satellite data from various sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agesofman/cronus}
  \item \url{https://agesofman.github.io/cronus/}
  \item Report bugs at \url{https://github.com/agesofman/cronus/issues}
}

}
\author{
\strong{Maintainer}: Ioannis Oikonomidis \email{goikon@math.uoa.gr} (\href{https://orcid.org/0000-0001-8130-2104}{ORCID})

Authors:
\itemize{
  \item Samis Trevezas \email{strevezas@math.uoa.gr} [thesis advisor]
}

}
\keyword{internal}
