% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_classes.R
\docType{class}
\name{Database-class}
\alias{Database-class}
\alias{Database}
\alias{Sector-class}
\alias{Sector}
\alias{Provider-class}
\alias{Provider}
\alias{Product-class}
\alias{Product}
\alias{Variable-class}
\alias{Variable}
\alias{Agricultural-class}
\alias{Agricultural}
\alias{Environmental-class}
\alias{Environmental}
\alias{Satellite-class}
\alias{Satellite}
\alias{Nass-class}
\alias{Nass}
\alias{Bibliography-class}
\alias{Bibliography}
\alias{Ornl-class}
\alias{Ornl}
\alias{Modis-class}
\alias{Modis}
\alias{Landsat-class}
\alias{Landsat}
\alias{Sentinel-class}
\alias{Sentinel}
\alias{Quickstats-class}
\alias{Quickstats}
\alias{Cropmaps-class}
\alias{Cropmaps}
\alias{Parameters-class}
\alias{Parameters}
\alias{Daymet-class}
\alias{Daymet}
\alias{Mod09ga-class}
\alias{Mod09ga}
\alias{Myd09ga-class}
\alias{Myd09ga}
\alias{Landsat_tm_c1-class}
\alias{Landsat_tm_c1}
\alias{Landsat_etm_c1-class}
\alias{Landsat_etm_c1}
\alias{Landsat_8_c1-class}
\alias{Landsat_8_c1}
\alias{S2msi2a-class}
\alias{S2msi2a}
\alias{Sy_2_syn___-class}
\alias{Sy_2_syn___}
\title{Cronus database classes}
\value{
An S4 object of the appropriate class (constructor functions).
}
\description{
cronus creates a database, which takes the form of a 5-level directory. These
levels are Region > Sector > Provider > Product > Variable.
}
\details{
For a specific region, data can belong to one of three sectors: agricultural,
environmental or satellite. Data are made available by a provider, usually an
organization or a lab, which organizes the data in one or more products.
Each product contains several variables of interest.

The class tree structure is the following:
\itemize{
\item \code{Agricultural} Sector
\itemize{
\item \code{Nass} Provider
\itemize{
\item \code{Quickstats} Product
\itemize{
\item \code{Progress} Variable
}
\item \code{Cropmaps} Product
}
\item \code{Bibliography} Provider
\itemize{
\item \code{Parameters} Product
}
}
\item \code{Environmental} Sector
\itemize{
\item \code{Ornl} Provider
\itemize{
\item \code{Daymet} Product
}
}
\item \code{Satellite} Sector
\itemize{
\item \code{Modis} Provider
\itemize{
\item \code{Mod09ga} Product
\item \code{Myd09ga} Product
}
\item \code{Landsat} Provider
\itemize{
\item \code{Landsat_tm_c1} Product
\item \code{Landsat_etm_c1} Product
\item \code{Landsat_8_c1} Product
}
\item \code{Sentinel} Provider
\itemize{
\item \code{S2msi2a} Product
\item \code{Sy_2_syn___} Product
}
}
}

Classes \code{Database}, \code{Sector}, \code{Provider}, and \code{Product}
are also defined, but are abstract classes that exist solely for hierarchical
purposes.

All classes are defined with the S4 OOP system, except from the Variables
that are defined using the S3 OOP system. Each variable has its own
documentation page.
}
\section{Slots}{

\describe{
\item{\code{region}}{Region. A region of interest.}

\item{\code{date}}{Date. Dates of interest (vector).}

\item{\code{dir}}{character. The database parent directory. If the path_demeter
environment variable has been set, it can be omitted.}
}}

\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- as.Date("2020-07-15")
dir <- getwd()

w <- new("Quickstats", region = region, date = date, dir = dir)
x <- new("Cropmaps", region = region, date = date, dir = dir)
y <- new("Daymet", region = region, date = date, dir = dir)
z <- new("Mod09ga", region = region, date = date, dir = dir)

}
}
\seealso{
\code{\link[=set_path_demeter]{set_path_demeter()}}
}
