% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\alias{derive,Daymet-method}
\alias{derive,Satellite-method}
\title{Derive data}
\usage{
derive(x, ...)

\S4method{derive}{Daymet}(x, y, variable, varxy, ...)

\S4method{derive}{Satellite}(x, variable)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments to \code{variable}.}

\item{y}{S4 object. A product of interest.}

\item{variable}{character. A function to compute the variable of interest.}

\item{varxy}{character. Variables used to derive \code{variable}.}
}
\value{
nothing. The data are saved directly in the cronus database.
}
\description{
Derive variables based on existing data.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

## Daymet

# Create objects
x <- new("Daymet", region = region, date = date)
y <- new("Cropmaps", region = region, date = date)

# Create parameters (cardinal temperatures)
z <- new("Parameters", region = region, date = date)
tb_ct <- read(z, "default")$tb_ct

# Derive gdd (take a look at the gdd function)
derive(x, y, "gdd", varxy = c("tmin", "tmax", "cdl_default"), tb_ct = tb_ct)

## MOD09GA

# Create objects
x <- new("Mod09ga", region = region, date = date)

# Derive ndvi and cloudmask
derive(x, "ndvi")
derive(x, "cloudmask")
}
}
