% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{download,Quickstats-method}
\alias{download,Cropmaps-method}
\alias{download,Daymet-method}
\alias{download,Satellite-method}
\title{Download data}
\usage{
download(x, ...)

\S4method{download}{Quickstats}(x, variable, ringname = NULL)

\S4method{download}{Cropmaps}(x, variable = NULL)

\S4method{download}{Daymet}(x, variable = NULL)

\S4method{download}{Satellite}(x, ringname = NULL)
}
\arguments{
\item{x}{S4 object. The product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. The variable of interest.}

\item{ringname}{character. A keyring to handle the provider credentials.}
}
\value{
For most products, nothing is returned. The data are saved directly in the cronus database. For lightweight data (such as variables of the quickstats product), the downloaded data are returned (as a data.frame).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Download agricultural, environmental, or satellite data for a region and time of interest.
}
\details{
Currently, the function can download:
\itemize{
\item The Quickstats product of NASS.
\item The Cropland Data Layer variable of NASS.
\item The Daymet product of ORNL.
\item The MOD09GA and MYD09GA products of MODIS.
\item The Landsat TM, ETM and 8 products.
\item The Sentinel-2 MSI and Sentinel-3 Synergy products.
}
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

# Create a keyring
ringname <- "my_ringname"
password <- "my_password"
create_keyring(ringname, password)
log_in(ringname, password)

# Add an API to download the quickstats data.
add_key(ringname = ringname,
        provider = "nass",
        username = NULL,
        password = "nass_key")

# Download the progress variable from the Quickstats product.
x <- new("Quickstats", region = region, date = date, dir = dir)
variable <- "progress"
data <- download(x, variable, ringname)

# Download the Cropland Data Layer (CDL) variable from the Cropmaps product.
x <- new("Cropmaps", region = region, date = date, dir = dir)
variable <- "cdl"
download(x, variable)

log_out(ringname)

}
}
