% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermes.R
\name{download_hermes}
\alias{download_hermes}
\title{Download project \code{hermes}}
\usage{
download_hermes(dir = getwd())
}
\arguments{
\item{dir}{character. The parent directory inside which the project will
be downloaded}
}
\value{
nothing. The project is downloaded in the specified directory.
}
\description{
Download project \code{hermes} from the github repository.
}
\examples{
\dontrun{
# Set path
path_hermes <- getwd()

# Download project
download_hermes(path_hermes)

# Create files object
files <- PersephoneFiles(prm = "test", dts = "test")

# Create model object
object <- create(files)
}
}
