% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{dtri}
\alias{dtri}
\title{Triangular distribution density}
\usage{
dtri(x, min = 0, max = 1, mode = 0.5)
}
\arguments{
\item{x}{numeric. A vector of values.}

\item{min}{numeric. The support minimum.}

\item{max}{numeric. The support maximum.}

\item{mode}{numeric. The mode (argmax) of the density function.}
}
\value{
A vector of the same length as \code{x}.
}
\description{
Calculate the triangular density function value.
}
\examples{
\dontrun{
x <- c(0.2, 0.5, -2, 1.5, NA)
dtri(x, min = 0, max = 1, mode = 0.5)
}
}
