% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rprofile.R
\name{edit_prof_line}
\alias{edit_prof_line}
\alias{edit_prof_options}
\title{Edit .RProfile}
\usage{
edit_prof_line(newline, prof = "user")

edit_prof_options(variable, value, prof = "user")
}
\arguments{
\item{newline}{character. Line to add in the file.}

\item{prof}{character. The type of profile, one of 'user' or 'project'.}

\item{variable}{character. Variable to add in \code{options(variable = value)}.}

\item{value}{character. Value to add in \code{options(variable = value)}..}
}
\value{
nothing. The .RProfile file is edited.
}
\description{
These functions can help the user edit an .RProfile file
without actually opening the file.
}
\section{Functions}{
\itemize{
\item \code{edit_prof_line()}: Add a line of code in the .RProfile file.

\item \code{edit_prof_options()}: Add a call to \code{options()} in the .RProfile file.

}}
\examples{
\dontrun{
edit_prof_line("library(agesofman)", prof = "user")
edit_prof_options("path_demeter", getwd(), prof = "user")
}
}
\seealso{
\code{\link[=set_path_demeter]{set_path_demeter()}}
}
