% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillgaps.R
\name{fillgaps}
\alias{fillgaps}
\alias{fillgaps,Satellite-method}
\title{Fill gaps}
\usage{
fillgaps(x, ...)

\S4method{fillgaps}{Satellite}(x, variable)
}
\arguments{
\item{x}{S4 object. The product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. The variable of interest.}
}
\value{
nothing. The data are saved directly in the cronus database.
}
\description{
Find dates for which the raster is missing and fill the gap with an empty
raster.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

## MOD09GA

# Create objects
x <- new("Mod09ga", region = region, date = date)

# Create the NDVI rasters that are missing
fillgaps(x, "ndvi")

# Create the cloud mask rasters that are missing
fillgaps(x, "cloudmask")
}
}
