% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{gdd}
\alias{gdd}
\title{Compute Growing Degree Days (GDD)}
\usage{
gdd(x, y, a, tb_ct)
}
\arguments{
\item{x}{numeric. Minimum temperature.}

\item{y}{numeric. Maximum temperature.}

\item{a}{character. Crop name.}

\item{tb_ct}{matrix. The metadata storing the crop cardinal temperatures.}
}
\value{
A vector with the same dimensions as \code{x} and \code{y}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Compute a common expression of thermal time, the Growing Degree Days (GDD).
}
