% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_branch}
\alias{get_branch}
\alias{create_db}
\alias{print_tree}
\title{Manage database}
\usage{
get_branch(
  dir = NULL,
  region = NULL,
  sector = NULL,
  provider = NULL,
  product = NULL,
  variable = NULL
)

create_db(
  dir = NULL,
  region = NULL,
  sector = NULL,
  provider = NULL,
  product = NULL,
  variable = NULL
)

print_tree()
}
\arguments{
\item{dir}{character. The database parent directory.}

\item{region}{Region The region of interest.}

\item{sector}{character. The sector of interest.}

\item{provider}{character. The provider of interest.}

\item{product}{character. The product of interest.}

\item{variable}{character. The variable of interest.}
}
\value{
\itemize{
\item \code{get_branch()} returns a list containing the \code{path} (absolute or relative,
depending on whether the parent directory was provided or not), as well as
all the arguments provided and those that could be extrapolated from them.
\item \code{create_db()} creates the directory specified, \code{get_branch()$path}, and
returns the path invisibly.
\item \code{print_tree()} prints the database structure in a Sector - Provider - Product
tree form.
}
}
\description{
Get the relative or absolute path to a branch of the \code{cronus} database.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

list_cropmaps <- get_branch(dir = getwd(), region = region, product = "cropmaps")
list_cropmaps$dir
list_cropmaps$region
list_cropmaps$sector
list_cropmaps$provider
list_cropmaps$product
list_cropmaps$variable
list_cropmaps$path

path_cropmaps <- create_db(dir = getwd(), region = region, product = "cropmaps")
}
}
