% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_files}
\alias{get_files}
\alias{get_files,Mod09ga,ANY-method}
\alias{get_files,Daymet,Cropmaps-method}
\title{Get files}
\usage{
get_files(x, y, ...)

\S4method{get_files}{Mod09ga,ANY}(x, variable)

\S4method{get_files}{Daymet,Cropmaps}(x, y, variable, varxy)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{y}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. The variable of interest.}

\item{varxy}{character. Variables used to derive \code{variable}.}
}
\value{
A vector or list with the paths required by \code{derive}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Get the file paths needed to derive a variable.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

x <- new("Mod09ga", region = region, date = date, dir = dir)
get_files(x, "ndvi")
}
}
