% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_level}
\alias{get_level}
\alias{is_sector}
\alias{is_provider}
\alias{is_product}
\alias{get_class}
\alias{get_sector}
\alias{get_provider}
\alias{get_product}
\title{Manage database}
\usage{
get_level(name)

is_sector(name)

is_provider(name)

is_product(name)

get_class(x)

get_sector(x)

get_provider(x)

get_product(x)
}
\arguments{
\item{name}{character. A class name of interest.}

\item{x}{S4 object. An object of a class defined in \code{cronus}.}
}
\value{
\itemize{
\item \code{get_level} returns a character, the level name of the structured database.
\item \code{is_x} returns a logical. It answers the question "Is name a x?"
\item \code{get_x} returns a character, the name of the object's x.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Helper functions concerning the database structure.
}
\examples{
\dontrun{
get_level("modis")
get_level("cdl")

is_sector("Agricultural")
is_provider("cdl")
}
}
