% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rprofile.R
\name{get_path_prof}
\alias{get_path_prof}
\title{Get .RProfile path}
\usage{
get_path_prof(prof)
}
\arguments{
\item{prof}{character. The type of profile, one of 'user' or 'project'.}
}
\value{
character. The .RProfile file path.
}
\description{
.RProfile is a file loaded at the startup of each R session.
It can be modified to contain code, so that the user does not have to
explicitely execute the same commands every single time.
}
\details{
There are multiple locations in which .RProfile files are stored.
This function can only modify the 'user' and the 'project' .RProfile files.
Here, 'project' refers to an RStudio project. The examples show all possible
locations of the .RProfile files. See also \link[base]{Startup}.
}
\examples{
\dontrun{
get_path_prof("user")
get_path_prof("project")

# Locations of .RProfile files in loading order
Sys.getenv("R_PROFILE")
file.path(Sys.getenv("R_HOME"), "etc", "Rprofile.site")
Sys.getenv("R_PROFILE_USER")
file.path(getwd(), ".Rprofile")
file.path(Sys.getenv("HOME"), ".Rprofile") # this is selected by 'prof = user'.
}
}
\seealso{
\link[base:Startup]{base::Startup}, \code{\link[rprojroot:find_root_file]{rprojroot::find_rstudio_root_file()}}
}
