% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{get_roi}
\alias{get_roi}
\alias{get_roi,Region-method}
\title{Get a region of interest (roi)}
\usage{
get_roi(region, ...)

\S4method{get_roi}{Region}(region)
}
\arguments{
\item{region}{character. The region of interest. See details}

\item{...}{extra arguments}
}
\value{
An object of class sf.
}
\description{
This function takes the name of a region and returns a sf object.
}
\details{
The function currently works only with US states and counties. The
object is retrieved from the U.S. Census TIGER shapefiles. Argument
\code{region} takes the form of a named character vector with elements:
\itemize{
\item \code{name}, the name of the region.
\item \code{type}, currently supporting "US State" and "US County".
\item \code{state}, the name of the US State. Used only for US Counties.
}
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
get_roi(region)

region <- Region(name = "nance", type = "us county",
                 div = c(country = "United States", state = "Nebraska", county = "Nance"))
get_roi(region)
}
}
