% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demeter.R
\name{get_service}
\alias{get_service}
\title{Provider service}
\usage{
get_service(provider)
}
\arguments{
\item{provider}{character. A provider of interest.}
}
\value{
character. The name of the corresponding service.
}
\description{
Get the service that keeps the satellite data stored in order to access the data.
}
\examples{
\dontrun{
get_service("modis")
get_service("sentinel")
}
}
