% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toi.R
\name{get_toi}
\alias{get_toi}
\alias{date_seq}
\alias{doy_to_date}
\alias{pdoy_to_date}
\title{Manage time of interest (\code{toi})}
\usage{
get_toi(date)

date_seq(from, to, by = "day")

doy_to_date(doy, year)

pdoy_to_date(pdoy)
}
\arguments{
\item{date}{character. A vector of dates.}

\item{from}{character. First date.}

\item{to}{character. Last date.}

\item{by}{character. Increment of sequence, e.g. day}

\item{doy}{numeric. A vector of days-of-year.}

\item{year}{numeric. A vector of years.}

\item{pdoy}{character. A vector of three-digit day-of-year (padded with zeros).}
}
\value{
\itemize{
\item \code{get_toi()} returns a list with elements \code{date}, \code{year}, \code{uyear}, \code{month}, \code{doy}, \code{pdoy}, and \code{name}.
\item \code{date_seq()}, \code{doy_to_date()}, \code{pdoy_to_date()} return a sequence of Dates.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Create a list that contains all the necessary formats of the time of interest.
}
\examples{
\dontrun{
dates <- date_seq("2002-01-01", "2002-01-04")
toi <- get_toi(dates)
doy_to_date(c(1, 365), c(2002, 2005))
pdoy_to_date(c("2002001", "2005365"))
}
}
