% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\alias{mask,Satellite-method}
\title{Mask rasters}
\usage{
mask(x, ...)

\S4method{mask}{Satellite}(x, variable1, variable2 = "cloudmask")
}
\arguments{
\item{x}{S4 object. The product of interest.}

\item{...}{extra arguments.}

\item{variable1}{character. The variable to be masked.}

\item{variable2}{character. The variable to be used as a mask.}
}
\value{
nothing. The data are saved directly in the cronus database.
}
\description{
Mask a raster variable using another raster variable as the mask.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

## MOD09GA

# Create objects
x <- new("Mod09ga", region = region, date = date)

# Apply a cloud mask on NDVI
mask(x, "ndvi", "cloudmask")
}
}
