% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify}
\alias{modify}
\alias{modify,Progress-method}
\alias{modify,ProgressList-method}
\title{modify data}
\usage{
modify(data, ...)

\S4method{modify}{Progress}(data, vars, fun = cumsum)

\S4method{modify}{ProgressList}(data, vars, fun = cumsum)
}
\arguments{
\item{data}{data.frame. The data to modify.}

\item{...}{extra arguments.}

\item{vars}{character. The name of the variables (columns) of \code{data}.}

\item{fun}{function. The name of a function to apply on \code{data}.}
}
\value{
An object of the same class as \code{data}.
}
\description{
modify data
}
