% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\alias{mutate,Progress-method}
\alias{mutate,ProgressList-method}
\title{Mutate data}
\usage{
mutate(data, ...)

\S4method{mutate}{Progress}(data, vars, fun = cumsum)

\S4method{mutate}{ProgressList}(data, vars, fun = cumsum)
}
\arguments{
\item{data}{data.frame. The data to mutate.}

\item{...}{extra arguments.}

\item{vars}{character. The name of the variables (columns) of \code{data}.}

\item{fun}{function. The name of a function to apply on \code{data}.}
}
\value{
An object of the same class as \code{data}.
}
\description{
Mutate data
}
