% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ndvi}
\alias{ndvi}
\title{Compute MODIS Normalized Difference Vegetation Index (NDVI)}
\usage{
ndvi(x, y)
}
\arguments{
\item{x}{numeric. Near-Infrared reflectance.}

\item{y}{numeric. Red reflectance.}
}
\value{
A vector with the same dimensions as \code{x} and \code{y}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Compute MODIS Normalized Difference Vegetation Index (NDVI)
}
