% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_progress.R
\name{new_Progress}
\alias{new_Progress}
\alias{validate_Progress}
\alias{Progress}
\alias{new_ProgressList}
\alias{validate_ProgressList}
\alias{ProgressList}
\alias{[.ProgressList}
\title{Progress Class}
\usage{
new_Progress(x = NULL)

validate_Progress(x)

Progress(x = NULL)

new_ProgressList(x = NULL)

validate_ProgressList(x)

ProgressList(x = NULL)

\method{[}{ProgressList}(x, i)
}
\arguments{
\item{x}{data.frame. The object to gain the class of interest. See Details.}

\item{i}{integer. The index passed in \code{[}.}
}
\value{
An object of class \code{Progress} or \code{ProgressList}.
}
\description{
An S3 class that contains the necessary information to study
the large-scale progress of the crops, expressed as the crop percentage that
occupies each phenological stage over an area of interest.
}
\details{
The \code{data.frame()} must include the following columns:
\itemize{
\item Crop (\code{character})
\item Stage (\code{factor})
\item Season (\code{numeric})
\item Time (\code{numeric})
\item Date (\code{Date})
\item Percentage (\code{numeric})
\item CumPercentage (\code{numeric})
}
}
\examples{
\dontrun{
# Simple initialization
x <- Progress()
class(x)
x <- ProgressList()
class(x)
class(x[[1]])

# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

# Quickstats Progress

# Create the object
x <- new("Quickstats", region = region, date = date)

# Download the data
data <- download(x, "progress", ringname)
class(data)
class(data[["Corn"]])
head(data$Corn)

# Plot the data
plot(data, crops = "Winter Wheat", year = 2021)
}
}
