% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,Cropmaps,character-method}
\alias{plot,Cropmaps,character-method}
\alias{plot,Daymet,character-method}
\title{Plot Rasters}
\usage{
\S4method{plot}{Cropmaps,character}(x, y, crops = NULL, year = NULL)

\S4method{plot}{Daymet,character}(x, y, date = NULL)
}
\arguments{
\item{x}{an S4 object. A product of interest.}

\item{y}{character. A variable of interest.}

\item{crops}{character. The crops of interest.}

\item{year}{numeric. Year(s) of interest.}

\item{date}{character Date(s) of interest.}
}
\value{
A plot, created with \code{terra}.
}
\description{
Create a plot of the data stored in the database.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

# Create the object
x <- new("Cropmaps", region = region, date = date)

# Plot
variable <- "cdl"
plot(x, variable, crops = c("Soybeans", "Winter Wheat"), year = 2002)

## Daymet Tmin

# Create the object
x <- new("Daymet", region = region, date = date, dir = dir)

# Plot
plot(x, "tmin", "2002-01-01")
}
}
