% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,Progress,ANY-method}
\alias{plot,Progress,ANY-method}
\title{Plot Progress}
\usage{
\S4method{plot}{Progress,ANY}(
  x,
  season = NULL,
  crops = NULL,
  ncol = NULL,
  nrow = NULL,
  ylab = "Percentage",
  xlab = "Time",
  save = FALSE,
  dir = getwd(),
  file = "plot.pdf",
  width = 15,
  height = 8
)
}
\arguments{
\item{x}{an object of class \code{Progress}.}

\item{season}{numeric. Year(s) of interest.}

\item{crops}{character. The crops of interest.}

\item{ncol}{numeric. Number of rows in the plot layout.}

\item{nrow}{numeric. Number of columns in the plot layout.}

\item{ylab}{character. The y-axis title.}

\item{xlab}{character. The x-axis title.}

\item{save}{logical. Should the plot be saved?}

\item{dir}{character. The directory in which the plot will be saved.}

\item{file}{character. The file name.}

\item{width}{numeric. The width of the plot in inches.}

\item{height}{numeric. The height of the plot in inches.}
}
\value{
A plot, created with \code{ggplot2}.
}
\description{
Create a plot of the crop progress data.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

# Create the object
x <- new("Quickstats", region = region, date = date)

# Download the data
data <- download(x, "progress", ringname)
class(data)
class(data[["Corn"]])
head(data$Corn)

# Plot the data
plot(data, crops = "Winter Wheat", year = 2021)
}
}
