% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot,Quickstats,character-method}
\alias{plot,Cropmaps,character-method}
\alias{plot,Daymet,character-method}
\title{Plot}
\usage{
plot(x, y, ...)

\S4method{plot}{Quickstats,character}(
  x,
  y,
  year,
  crops = NULL,
  ncol = NULL,
  nrow = NULL,
  ylab = "Percentage",
  xlab = "Time",
  save = FALSE,
  dir = getwd(),
  file = "plot.pdf",
  width = 15,
  height = 8
)

\S4method{plot}{Cropmaps,character}(x, y, crops = NULL, year = NULL)

\S4method{plot}{Daymet,character}(x, y, date)
}
\arguments{
\item{x}{an S4 object. A product of interest.}

\item{y}{character. A variable of interest.}

\item{...}{extra arguments.}

\item{year}{numeric. Year(s) of interest.}

\item{crops}{character. The crops of interest.}

\item{ncol}{numeric. Number of rows in the plot layout.}

\item{nrow}{numeric. Number of columns in the plot layout.}

\item{ylab}{character. The y-axis title.}

\item{xlab}{character. The x-axis title.}

\item{save}{logical. Should the plot be saved?}

\item{dir}{character. The directory in which the plot will be saved.}

\item{file}{character. The file name.}

\item{width}{numeric. The width of the plot in inches.}

\item{height}{numeric. The height of the plot in inches.}

\item{date}{character Date(s) of interest.}
}
\value{
A plot, created either with \code{ggplot2} or with \code{terra}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Create a plot of the data stored in the database.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

# Plot the Cropland Data Layer
x <- new("Cropmaps", region = region, date = date, dir = dir)
variable <- "cdl"
download(x, variable)
plot(x, variable, crops = c("Soybeans", "Winter Wheat"), year = 2002)

# Plot the Daymet minimum temperature
x <- new("Daymet", region = region, date = date, dir = dir)
download(x, variable = "tmin")
plot(x, "tmin", "2002-01-01")
}
}
