% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\alias{project,Cropmaps,Daymet-method}
\title{Project rasters}
\usage{
project(x, y, ...)

\S4method{project}{Cropmaps,Daymet}(x, y, variablex, variabley, newvarname = variablex)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{y}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{variablex}{character. The variable of product \code{x}.}

\item{variabley}{character. The variable of product \code{y}.}

\item{newvarname}{character. The name of the projected variable.}
}
\value{
nothing. The data are saved directly in the cronus database.
}
\description{
Project a raster variable to match another raster variable.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

## Cropmaps CDL and Daymet

# Create objects
x <- new("Cropmaps", region = region, date = date)
y <- new("Daymet", region = region, date = date)

# Project
project(x, y, variablex = "cdl", variabley = "tmin", "cdl_projected")
}
}
