% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermes.R
\name{read_exe}
\alias{read_exe}
\title{Read scripts}
\usage{
read_exe(name, project, dir = get_path_hermes())
}
\arguments{
\item{name}{character. The name of the script (without the .R suffix).}

\item{project}{character. The project the script concerns.}

\item{dir}{character. The hermes project directory.}
}
\description{
This set of functions sources scripts inside the structured
directory of the \code{hermes} project and returns an appropriate object. These
functions are used in the model creation process.
}
\section{Functions}{
\itemize{
\item \code{read_exe()}: source the script.

}}
\examples{
\dontrun{
# Set path
path_hermes <- getwd()

# Download project
download_hermes(path_hermes)

# Read files
read_exe("test", "persephone", dir = path_hermes)
prm <- read_prm("test", "persephone", dir = path_hermes)
}
}
