% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\alias{recode,Cropmaps-method}
\title{Recode data}
\usage{
recode(x, ...)

\S4method{recode}{Cropmaps}(x, variable, mdname, varname = variable)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. A function to compute the variable of interest.}

\item{mdname}{character. The metadata file names needed for the recoding.}

\item{varname}{character. The name of the recoded variable.}
}
\value{
nothing. The data are saved directly in the cronus database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Recode variables using metadata.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

x <- new("Cropmaps", region = region, date = date, dir = dir)
recode(x, "cdl_default", mdname = "default")
}
}
