% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.R
\name{refactor}
\alias{refactor}
\alias{refactor,Cropmaps-method}
\title{Refactor data}
\usage{
refactor(x, ...)

\S4method{refactor}{Cropmaps}(x, variable, mdname, newvarname = variable)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. The variable of interest.}

\item{mdname}{character. The metadata file names needed for the recoding.}

\item{newvarname}{character. The name of the refactored variable.}
}
\value{
nothing. The data are saved directly in the cronus database.
}
\description{
Refactor variables using metadata.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

## Cropmaps CDL

# Create the object
x <- new("Cropmaps", region = region, date = date)

# Get the current categories data.frame
df_cat <- terra::cats(cdl_2020)[[1]]

# Define a transformation table
tb_rcl <- cbind(a = c(121:124, 141:143), b = c(rep(82, 4), rep(63, 3)))

# Write the metadata
write(x, name = "my_metadata", df_cat = df_cat, tb_rcl = tb_rcl)

# Refactor the raster
refactor(x, variable = "cdl_projected", mdname = "my_metadata", newvarname = "cdl_refactored")
}
}
