% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Region-class}
\alias{Region-class}
\alias{Region}
\title{Region class}
\value{
an object of class \code{Region}.
}
\description{
An S4 class that defines a region of interest.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. The name of the region.}

\item{\code{type}}{character. The type of the region.}

\item{\code{id}}{ANY. An appropriate ID for the region.}

\item{\code{div}}{character. A named character vector specifying the divisions,
starting from the country level and stopping at the region's level.}
}}

\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

region <- Region(name = "nance", type = "us county",
                 div = c(country = "United States", state = "Nebraska", county = "Nance"))
}
}
