% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermes.R
\name{save_hermes}
\alias{save_hermes}
\alias{read_hermes}
\title{Save to and read from hermes}
\usage{
save_hermes(object, ...)

read_hermes(object, ...)
}
\arguments{
\item{object}{ANY. The object to save / read.}

\item{...}{extra arguments}
}
\value{
\code{save_hermes()} returns the absolute path in which the object is
saved is returned invisibly. \code{read_hermes()} returns the object of interest.
}
\description{
Save an object to the \code{hermes} database with \code{save_hermes()} and easily
retrieve it with \code{read_hermes()}.
}
