% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rprofile.R
\name{set_path_demeter}
\alias{set_path_demeter}
\alias{get_path_demeter}
\title{Set Demeter path}
\usage{
set_path_demeter(path, prof = "user")

get_path_demeter()
}
\arguments{
\item{path}{character. The path to the database directory.}

\item{prof}{character. The type of profile, one of 'user' or 'project'.}
}
\value{
nothing. The .RProfile file is edited.
}
\description{
Set a default path_demeter by editing the .RProfile file.
}
\examples{
\dontrun{
set_path_demeter(getwd(), prof = "user")
get_path_demeter()
}
}
\seealso{
\code{\link[=edit_prof_options]{edit_prof_options()}}
}
