% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_startup.R
\name{set_path_hermes}
\alias{set_path_hermes}
\alias{get_path_hermes}
\title{Set hermes path}
\usage{
set_path_hermes(path)

get_path_hermes()
}
\arguments{
\item{path}{character. The path to the database directory.}
}
\value{
nothing. The .Renviron file is edited.
}
\description{
Set or get the default \code{path_hermes} by editing the .Renviron
file.
}
\examples{
\dontrun{
set_path_hermes(getwd())
get_path_hermes()
}
}
\seealso{
\code{\link[=get_startup_path]{get_startup_path()}}
}
