% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothout.R
\name{smoothout}
\alias{smoothout}
\alias{smoothout,Satellite-method}
\title{Smooth raster time-series}
\usage{
smoothout(x, ...)

\S4method{smoothout}{Satellite}(x, variable, smoother = "wt1", pre = NULL, post = NULL)
}
\arguments{
\item{x}{S4 object. The product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. The variable of interest.}

\item{smoother}{character. A function to be applied on the original data.}

\item{pre}{character. A vector of dates to be used as an extra smoothing window before the dates of interest.}

\item{post}{character. A vector of dates to be used as an extra smoothing window after the dates of interest.}
}
\value{
nothing. The smoothed data are saved directly in the cronus database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Smooth raster time-series and interpolate missing values using a smoother function.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

x <- new("Mod09ga", region = region, date = date, dir = dir)
pre <- NULL
post <- NULL
smoothout(x, "ndvi", pre = pre, post = post)
}
}
