% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{summarize,Cropmaps-method}
\title{Summarize raster data}
\usage{
summarize(x, ...)

\S4method{summarize}{Cropmaps}(x, variable)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. A function to compute the variable of interest.}
}
\value{
data.frame. A data frame with the summarized data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize raster data into a data frame.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

x <- new("Cropmaps", region = region, date = date, dir = dir)
mydf <- summarize(x, "cdl_default")
}
}
