% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{tidy_progress}
\alias{tidy_progress}
\alias{calc_stages}
\alias{calc_percentage}
\title{Handle Quickstats progress data}
\usage{
tidy_progress(data)

calc_stages(data)

calc_percentage(data, cum = TRUE)
}
\arguments{
\item{data}{data.frame or list. The quickstats progress variable.}

\item{cum}{logical. Calculate the cumulative percentage? If FALSE,
calculate the simple percentage.}
}
\value{
data.frame or list.
}
\description{
Tidy the Quickstats progress data.
}
\details{
This function relies on the \code{tidyverse} package to tidy the data.
}
