% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_Qs_progress}
\alias{tidy_Qs_progress}
\alias{calc_cumperc}
\alias{calc_cumperc.Progress}
\alias{calc_cumperc.ProgressList}
\alias{calc_perc}
\alias{calc_perc.Progress}
\alias{calc_perc.ProgressList}
\title{Tidy Quickstats progress}
\usage{
tidy_Qs_progress(data)

calc_cumperc(data)

\method{calc_cumperc}{Progress}(data)

\method{calc_cumperc}{ProgressList}(data)

calc_perc(data)

\method{calc_perc}{Progress}(data)

\method{calc_perc}{ProgressList}(data)
}
\arguments{
\item{data}{data.frame. The quickstats progress variable.}
}
\value{
A \code{ProgressList}.
}
\description{
Tidy the Quickstats progress data.
}
\details{
This function relies on the \code{tidyverse} package to tidy the data.
}
