% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{trunc_minmax}
\alias{trunc_minmax}
\title{Truncate value}
\usage{
trunc_minmax(x, a, b)
}
\arguments{
\item{x}{numeric. A vector of values to be truncated.}

\item{a}{numeric. The lower truncation limit.}

\item{b}{numeric. The upper truncation limit.}
}
\value{
A vector with the same dimensions as \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Truncate the values of a vector between a lower and an upper limit.
}
\examples{
\dontrun{
x <- 1:10
trunc_minmax(x, 2, 7)
}
}
