% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\alias{write,Cropmaps-method}
\alias{write,Parameters-method}
\title{Write metadata}
\usage{
write(x, ...)

\S4method{write}{Cropmaps}(x, name, df_cat = NULL, tb_rcl = NULL)

\S4method{write}{Parameters}(x, name, tb_ct = NULL)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{name}{character. The name of the metadata files.}

\item{df_cat}{data.frame A data frame holding the cdl categories.}

\item{tb_rcl}{matrix A matrix holding the cdl value recoding.}

\item{tb_ct}{matrix. A matrix holding the cardinal temperatures.}
}
\value{
nothing. The data are saved directly in the cronus database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Write metadata files in the cronus database.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

## Parameters

# Create the object
x <- new("Parameters", region = region, date = date)

# Define the cardinal temperatures metadata
tb_ct <- rbind('Dry Beans' = c(Tb = 4.0,  To = 23.0, Tc = 32.0),
               'Corn'      = c(Tb = 10.0, To = 30.0, Tc = 50.0),
               'Soybeans'  = c(Tb = 6.0,  To = 26.0, Tc = 39.0))

# Write the metadata
write(y, name = "default", tb_ct = tb_ct)
}
}
