% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_metadata}
\alias{write_metadata}
\alias{write_metadata,Cropmaps-method}
\alias{write_metadata,Parameters-method}
\title{Write metadata}
\usage{
write_metadata(x, ...)

\S4method{write_metadata}{Cropmaps}(x, name, df_cat = NULL, tb_rcl = NULL)

\S4method{write_metadata}{Parameters}(x, name, tb_ct = NULL)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{name}{character. The name of the metadata files.}

\item{df_cat}{data.frame A data frame holding the cdl categories.}

\item{tb_rcl}{matrix A matrix holding the cdl value recoding.}

\item{tb_ct}{matrix. A matrix holding the cardinal temperatures.}
}
\value{
nothing. The data are saved directly in the cronus database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Write metadata files in the cronus database.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

x <- new("Parameters", region = region, date = date, dir = dir)
# tb_ct is a matrix with cardinal temperatures
write_metadata(x, name = "default", tb_ct = tb_ct)
}
}
