% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\alias{compose,Database,Cropmaps-method}
\title{Compose rasters by maps}
\usage{
compose(x, y, ...)

\S4method{compose}{Database,Cropmaps}(x, y, variablex, variabley, fun = "mean", ...)
}
\arguments{
\item{x}{S4 object. The product of interest.}

\item{y}{S4 object. The product determining the zones.}

\item{...}{extra arguments passed to \code{fun}.}

\item{variablex}{character. The variable of product \code{x}.}

\item{variabley}{character. The variable of product \code{y}.}

\item{fun}{character or function. The function to apply for the summaries.}
}
\value{
A data.frame with the zonal summaries. The data.frame is also saved
in the cronus database in .csv format.
}
\description{
Compose Rasters into zonal summaries, indicated by another raster.
}
\details{
Currently, \code{y} can only be a NASS Cropmap.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")

## Daymet

# Create objects
x <- new("Daymet", region = region, date = date)
y <- new("Cropmaps", region = region, date = date)

# Compose
a <- compose(x, y, variablex = "gdd",  variabley = "cdl_default", fun = "mean")
b <- compose(x, y, variablex = "prcp", variabley = "cdl_default", fun = "mean")

## MOD09GA

# Create objects
x <- new("Mod09ga", region = region, date = date)
y <- new("Cropmaps", region = region, date = date)

a <- compose(x, y, variablex = "ndvi_smooth_wt1", variabley = "cdl_recoded", fun = "mean")
}
}
