% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermes.R
\name{create_dir_hermes}
\alias{create_dir_hermes}
\alias{save_hermes}
\alias{read_hermes}
\title{Save and read objects}
\usage{
create_dir_hermes(
  region = NULL,
  project = NULL,
  type = NULL,
  class = NULL,
  name = NULL,
  suffix = ".RData",
  dir = get_path_hermes()
)

save_hermes(
  object,
  region,
  project,
  type,
  class,
  name,
  suffix = ".RData",
  dir = get_path_hermes()
)

read_hermes(
  region,
  project,
  type,
  class,
  name,
  suffix = ".RData",
  dir = get_path_hermes()
)
}
\arguments{
\item{region}{Region. A region of interest.}

\item{project}{character. The project the script concerns (e.g. cronus).}

\item{type}{character. The type of the object (e.g. plot, report).}

\item{class}{character. The class of the object.}

\item{name}{character. The name of the script (without the suffix).}

\item{suffix}{character. The name of the suffix (starting with a dot).}

\item{dir}{character. The hermes project directory.}

\item{object}{ANY. The object to save / load.}
}
\description{
This set of functions saves and reads files inside the \code{hermes}
structured directory.
}
\examples{
\dontrun{
# Define required variables
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

create_dir_hermes(region, "cronus", "plot", "Progress", "myplot", ".pdf")
}
}
