% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Database-class}
\alias{Database-class}
\alias{Database}
\alias{Sector-class}
\alias{Sector}
\alias{Provider-class}
\alias{Provider}
\alias{Product-class}
\alias{Product}
\alias{Agricultural-class}
\alias{Agricultural}
\alias{Environmental-class}
\alias{Environmental}
\alias{Satellite-class}
\alias{Satellite}
\alias{Nass-class}
\alias{Nass}
\alias{Bibliography-class}
\alias{Bibliography}
\alias{Ornl-class}
\alias{Ornl}
\alias{Modis-class}
\alias{Modis}
\alias{Landsat-class}
\alias{Landsat}
\alias{Sentinel-class}
\alias{Sentinel}
\alias{Quickstats-class}
\alias{Quickstats}
\alias{Cropmaps-class}
\alias{Cropmaps}
\alias{Parameters-class}
\alias{Parameters}
\alias{Daymet-class}
\alias{Daymet}
\alias{Mod09ga-class}
\alias{Mod09ga}
\alias{Myd09ga-class}
\alias{Myd09ga}
\alias{Landsat_tm_c1-class}
\alias{Landsat_tm_c1}
\alias{Landsat_etm_c1-class}
\alias{Landsat_etm_c1}
\alias{Landsat_8_c1-class}
\alias{Landsat_8_c1}
\alias{S2msi2a-class}
\alias{S2msi2a}
\alias{Sy_2_syn___-class}
\alias{Sy_2_syn___}
\title{Cronus database classes}
\value{
An S4 object of the appropriate class (constructor functions).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
cronus creates a database, which takes the form of a 5-level directory. These levels are Region > Sector > Provider > Product > Variable. For a specific region, data can belong to one of three sectors: agricultural, environmental or satellite. Data are made available by a provider, usually an organization or a lab, which organizes the data in one or more products. Each product contains several variables of interest.

The class tree structure is the following:
\itemize{
\item \code{Agricultural} Sector
\itemize{
\item \code{Nass} Provider
\itemize{
\item \code{Quickstats} Product
\item \code{Cropmaps} Product
}
\item \code{Bibliography} Provider
\itemize{
\item \code{Parameters} Product
}
}
\item \code{Environmental} Sector
\itemize{
\item \code{Ornl} Provider
\itemize{
\item \code{Daymet} Product
}
}
\item \code{Satellite} Sector
\itemize{
\item \code{Modis} Provider
\itemize{
\item \code{Mod09ga} Product
\item \code{Myd09ga} Product
}
\item \code{Landsat} Provider
\itemize{
\item \code{Landsat_tm_c1} Product
\item \code{Landsat_etm_c1} Product
\item \code{Landsat_8_c1} Product
}
\item \code{Sentinel} Provider
\itemize{
\item \code{S2msi2a} Product
\item \code{Sy_2_syn___} Product
}
}
}

Classes \code{Database}, \code{Sector}, \code{Provider}, and \code{Product} are also defined, but are abstract classes that exist solely for hierarchical purposes.
}
\section{Slots}{

\describe{
\item{\code{region}}{Region. A region of interest.}

\item{\code{date}}{Date. Dates of interest (vector).}

\item{\code{dir}}{character. The database parent directory.}
}}

\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2021-12-31")
dir <- get_path_demeter()

x <- new("Quickstats", region = region, date = date, dir = dir)
y <- Cropmaps(region = region, date = date, dir = dir)
}
}
