% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\alias{derive,Daymet-method}
\alias{derive,Satellite-method}
\title{Derive data}
\usage{
derive(x, ...)

\S4method{derive}{Daymet}(x, y, variable, varxy, ...)

\S4method{derive}{Satellite}(x, variable)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments to \code{variable}.}

\item{y}{S4 object. A product of interest.}

\item{variable}{character. A function to compute the variable of interest.}

\item{varxy}{character. Variables used to derive \code{variable}.}
}
\value{
nothing. The data are saved directly in the cronus database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Derive variables based on existing data.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

# Derive MODIS Terra NDVI and cloudmask variables
x <- new("Mod09ga", region = region, date = date, dir = dir)
derive(x, "ndvi")
derive(x, "cloudmask")
}
}
