% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_map}
\alias{load_map}
\alias{load_map,Product-method}
\alias{load_map,Cropmaps-method}
\title{Load rasters}
\usage{
load_map(x, ...)

\S4method{load_map}{Product}(x, variable, date = NULL)

\S4method{load_map}{Cropmaps}(x, variable, date = NULL)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. The variable of interest.}

\item{date}{character. Dates of interest.}
}
\value{
SpatRaster.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Load rasters stored in the cronus database.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-07-15", "2002-07-16")
dir <- getwd()

x <- new("Daymet", region = region, date = date, dir = dir)
variable <- "gdd"
load_map(x, variable)
load_map(x, variable, as.Date("2002-08-01"))
}
}
