% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\alias{read,Quickstats-method}
\alias{read,Cropmaps-method}
\alias{read,Parameters-method}
\alias{read,Daymet-method}
\alias{read,Mod09ga-method}
\alias{read,Database-method}
\title{Read data}
\usage{
read(x, ...)

\S4method{read}{Quickstats}(x, variable)

\S4method{read}{Cropmaps}(x, name)

\S4method{read}{Parameters}(x, name)

\S4method{read}{Daymet}(x, variable)

\S4method{read}{Mod09ga}(x, variable)

\S4method{read}{Database}(x, df_var)
}
\arguments{
\item{x}{S4 object. A product of interest.}

\item{...}{extra arguments.}

\item{variable}{character. The variable of interest.}

\item{name}{character. The name of the metadata files.}

\item{df_var}{data.frame. A data frame with columns Product and Variable.}
}
\value{
The data of interest (list, data.frame, raster, or other).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Read data stored in the cronus database.
}
\examples{
\dontrun{
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))
date <- date_seq("2002-01-01", "2002-12-31")
dir <- getwd()

x <- new("Quickstats", region = region, date = date, dir = dir)
variable <- "progress"
list_progress <- read(x, variable)

x <- new("Parameters", region = region, date = date, dir = dir)
read(x, name = "default")
}
}
