% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{rename_crops}
\alias{rename_crops}
\title{Rename crops}
\usage{
rename_crops(crop)
}
\arguments{
\item{crop}{character. A vector of crop names.}
}
\value{
A character vector of the same dimensions as \code{crops}.
}
\description{
Change specific crop names to avoid duplicates.
}
\examples{
\dontrun{
rename_crops(c("Corn", "Hay", "Wheat", "Beans"))
}
}
