% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_startup.R
\name{set_path_demeter}
\alias{set_path_demeter}
\alias{get_path_demeter}
\title{Set Demeter path}
\usage{
set_path_demeter(path, level = "user")

get_path_demeter()
}
\arguments{
\item{path}{character. The path to the database directory.}

\item{level}{character. The level of profile, one of 'user' or 'project'.}
}
\value{
nothing. The .RPenviron file is edited.
}
\description{
Set or get the default path_demeter by editing the .Renviron
file.
}
\examples{
\dontrun{
set_path_demeter(getwd(), level = "user")
get_path_demeter()
}
}
\seealso{
\code{\link[=get_startup_path]{get_startup_path()}}
}
